requirements.txt
line-bot-sdk==3.14.2
python-dotenv==1.0.1
pydantic==2.10.4
azure-ai-documentintelligence==1.0.0b4
gspread==6.1.4
oauth2client==4.1.3
boto3==1.35.90